pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clanbot"
version = "0.1.0"
description = "Telegram bot for Clash of Clans clan monitoring"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "OpenAI"}]
dependencies = [
  "aiogram>=3.26,<4.0",
  "SQLAlchemy>=2.0.36,<3.0",
  "aiosqlite>=0.20,<1.0",
  "alembic>=1.14,<2.0",
  "pydantic>=2.10,<3.0",
  "pydantic-settings>=2.6,<3.0",
  "APScheduler>=3.10.4,<4.0",
  "aiohttp>=3.10,<4.0",
  "PyYAML>=6.0,<7.0",
  "orjson>=3.10,<4.0",
  "tenacity>=9.0,<10.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<1.0",
  "pytest-cov>=5.0,<6.0",
  "freezegun>=1.5,<2.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]
